RAG_export/daily_planner_service/requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
chromadb
python-dotenv
uvicorn
fastapi
sentence-transformers
